--- .idea/misc.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="DesignSurface">
-    <option name="filePathToZoomLevelMap">
-      <map>
-        <entry key="..\:/KotlinProjects/ARTemplateProject/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.21875" />
-        <entry key="..\:/KotlinProjects/ARTemplateProject/app/src/main/res/drawable/bear.xml" value="0.21875" />
-        <entry key="..\:/KotlinProjects/ARTemplateProject/app/src/main/res/drawable/bg_round.xml" value="0.21875" />
-        <entry key="..\:/KotlinProjects/ARTemplateProject/app/src/main/res/drawable/ic_launcher_background.xml" value="0.21875" />
-        <entry key="..\:/KotlinProjects/ARTemplateProject/app/src/main/res/layout/activity_main.xml" value="0.28940217391304346" />
-        <entry key="..\:/KotlinProjects/ARTemplateProject/app/src/main/res/layout/example_widget.xml" value="0.23854166666666668" />
-      </map>
-    </option>
-  </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
-    <output url="file://$PROJECT_DIR$/build/classes" />
-  </component>
-  <component name="ProjectType">
-    <option name="id" value="Android" />
-  </component>
-</project>
