+++ app/build.gradle
@@ -13,7 +13,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.example.artemplateproject"
+        applicationId "com.example.synclient"
         minSdk 29
         targetSdk 31
         versionCode 1
