+++ app/src/androidTest/java/com/example/synclient/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.artemplateproject
+package com.example.synclient
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
