--- app/src/main/java/com/example/artemplateproject/common/helpers/SurfaceViewHelper.kt
@@ -1,4 +1,0 @@
-package com.example.artemplateproject.common.helpers
-
-object SurfaceViewHelper {
-}
