+++ app/src/main/java/com/example/synclient/common/helpers/CameraPermissionHelper.kt
