+++ app/src/main/java/com/example/synclient/common/helpers/SurfaceViewHelper.kt
@@ -1,0 +1,4 @@
+package com.example.synclient.common.helpers
+
+object SurfaceViewHelper {
+}
