+++ app/src/main/java/com/example/synclient/common/rendering/BackgroundRenderer.kt
