+++ app/src/main/java/com/example/synclient/common/rendering/PointCloudRenderer.kt
